requirements.txt
unsloth
torch
torchvision
torchaudio
transformers
trl
peft
accelerate
bitsandbytes
datasets
tqdm
pyyaml
python-dotenv
boto3
smart_open
chromadb
sentence-transformers
ipykernel
pandas
beautifulsoup4
requests
